refactor: split landing page into 8 focused components (#65)
Decompose 1093-line monolithic page.tsx into organized (landing)/ route group:
- Navigation.tsx: Fixed header with desktop/mobile menus
- HeroSection.tsx: Hero with mouse-follow glow and animated headline
- KanbanPreview.tsx: Interactive Kanban board with HTML5 D&D
- FeaturesSection.tsx: Feature cards with SSR-safe random subtitle
- HowItWorksSection.tsx: Four-step process cards
- CTASection.tsx: Call-to-action with center glow
- Footer.tsx: Simple footer with links
- Glow.tsx: Shared CVA-based decorative glow component

page.tsx reduced to ~30 lines of thin composition.
